requirements.txt
asgiref
Django
djangorestframework
python-dotenv
sqlparse
psycopg2-binary
